audit_log/requirements.txt
connexion==2.7.0
swagger-ui-bundle==0.0.8
requests==2.25.1
pykafka==2.4.0
sqlalchemy==1.3.24
starlette==0.16.0
flask_cors
